Streamline the predule file structure
