--- pkg/prelude/mutable_none.go
@@ -1,6 +1,0 @@
-package prelude
-
-// MutableNone instantiates an empty MutableOption.
-func MutableNone[T any]() OptionalMutable[T] {
-	return OptionalMutable[T]{nil}
-}
